// frontend/src/api/openapi/models/WordFrequencyRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WordFrequencyRead = {
  /**
   * ID of the SourceDocument
   */
  sdoc_id: number;
  /**
   * Word
   */
  word: string;
  /**
   * Count of the word in the SourceDocument
   */
  count: number;
};